requirements.txt
absl-py==1.2.0
asttokens==2.0.8
astunparse==1.6.3
attrs==22.1.0
Automat==20.2.0
backcall==0.2.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
constantly==15.1.0
contourpy==1.0.5
cryptography==38.0.1
cssselect==1.1.0
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.1.1
filelock==3.8.0
flatbuffers==22.9.24
fonttools==4.37.4
gast==0.4.0
google-auth==2.12.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.49.1
h5py==3.7.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==5.0.0
incremental==21.3.0
ipykernel==6.16.0
ipython==8.5.0
itemadapter==0.7.0
itemloaders==1.0.6
jedi==0.18.1
jmespath==1.0.1
jupyter-core==4.11.1
jupyter_client==7.4.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.3
oauthlib==3.2.1
opt-einsum==3.3.0
packaging==21.3
parsel==1.6.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
Protego==0.2.1
protobuf==3.19.6
psutil==5.9.2
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
Pygments==2.13.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pywin32==304
pyzmq==24.0.1
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.9
Scrapy==2.6.3
service-identity==21.1.0
six==1.16.0
stack-data==0.5.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
tldextract==3.4.0
tornado==6.2
traitlets==5.4.0
Twisted==22.8.0
twisted-iocpsupport==1.0.2
typing_extensions==4.4.0
urllib3==1.26.12
w3lib==2.0.1
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.9.0
zope.interface==5.5.0
